Remove duplicate error output on `abort`
There were two levels of duplication here.  Firstly, the abort
message was being printed to both stdout and stderr.  Secondly the
backstrace was bring printed twice.   I tested this change with both
node and with chrome and in both cases the stack is automatically
included in the string form of the exception.